+++ src/Kusto/Kusto.Autorest/test/Invoke-AzKustoDiagnoseClusterVirtualNetwork.Tests.ps1
@@ -16,7 +16,7 @@ Describe 'Invoke-AzKustoDiagnoseClusterVirtualNetwork' {
         . ($mockingPath | Select-Object -First 1).FullName
     }
     It 'Diagnose' -skip {
-        # Skiping this tests since this Operation is only supported for VNet injected clusters 
+        # Skipping this tests since this Operation is only supported for VNet injected clusters 
         $resourceGroupName = $env.resourceGroupName
         $clusterName = $env.kustoClusterName
 
@@ -24,7 +24,7 @@ Describe 'Invoke-AzKustoDiagnoseClusterVirtualNetwork' {
     }
 
     It 'DiagnoseViaIdentity' -skip {
-        # Skiping this tests since this Operation is only supported for VNet injected clusters 
+        # Skipping this tests since this Operation is only supported for VNet injected clusters 
         $resourceGroupName = $env.resourceGroupName
         $clusterName = $env.kustoClusterName