+++ src/LabServices/LabServices.Autorest/custom/Start-AzLabServicesUserVM_Start.ps1
@@ -47,7 +47,7 @@ function Start-AzLabServicesUserVM_Start {
         [Parameter(Mandatory)]
         [Microsoft.Azure.PowerShell.Cmdlets.LabServices.Category('Path')]
         [System.String]
-        # The name of the user that uniqely identifies it within containing lab.
+        # The name of the user that uniquely identifies it within containing lab.
         # Used in resource URIs.
         ${Email},