+++ src/Kusto/Kusto/ChangeLog.md
@@ -70,7 +70,7 @@
 
 ## Version 0.1.1
 * Capacity is a new and optional parameter for Create and Update Cluster.
-* ETag of Database is depricated.
+* ETag of Database is deprecated.
 * Cluster has new Properties: Uri, DataIngestionUri and Capacity.
 
 ## Version 0.1.0